+++ functions.php
@@ -34,6 +34,10 @@ add_filter( 'logout_redirect', 'hrs_logout_redirect_home', 10, 3 );
 add_filter( 'excerpt_length', 'hrs_excerpt_length' );
 add_filter( 'excerpt_more', 'hrs_excerpt_more_link' );
 
+// Plugin adjustments.
+add_filter( 'tablepress_table_render_data', 'hrs_add_hrs_filter_tablepress_atts', 10, 3 );
+add_action( 'after_setup_theme', 'hrs_gform_setup' );
+
 /**
  * Retrieves the version number from the main stylesheet headers.
  *
@@ -186,4 +190,183 @@ function hrs_logout_redirect_home( $redirect_to, $requested_redirect_to, $user )
 	$requested_redirect_to = home_url( '/' );
 
 	return $requested_redirect_to;
+}
+
+
+/**
+ * Adds TablePress attributes filter on tables with header rows.
+ *
+ * Callback function for the TablePress `tablepress_table_render_data`
+ * filter hook, called from the `TablePress_Render->_prepare_render_data()`
+ * method after processing the table visibility information and before
+ * rendering the table. This function simply checks whether the table has
+ * a designated header row and add a `tablepress_cell_tag_attributes` hook
+ * if it does. This prevents filtering the table cell attributes on tables
+ * without header rows.
+ *
+ * @since 1.0.0
+ *
+ * @param array $table          The processed table.
+ * @param array $orig_table     The unprocessed table.
+ * @param array $render_options The render options for the table.
+ * @return array The processed table to be rendered.
+ */
+function hrs_add_hrs_filter_tablepress_atts( $table, $orig_table, $render_options ) {
+	if ( true === $render_options['table_head'] ) {
+
+		// Clear any leftover data to fetch the lastest table contents.
+		delete_transient( 'hrs_tablepress_header_columns' );
+
+		// Filter each table cell.
+		add_filter( 'tablepress_cell_tag_attributes', 'hrs_filter_tablepress_atts', 10, 5 );
+
+	}
+
+	return $table;
+}
+
+/**
+ * Filters the TablePress cells to add `data-column` attributes.
+ *
+ * Adds a `data-column` attribute to every TablePress table cell with the
+ * value of the column header for that cell. This is used along with CSS to
+ * create a more responsive table layout. This method fires on every
+ * TablePress cell, so we save only the table properties we need (the
+ * contents of the header row and the numbers of the last row and column)
+ * in a WordPress transient and then delete it after the last cell is built.
+ * Called by the `tablepress_cell_tag_attributes` filter hook.
+ *
+ * @uses TablePress_Table_Model->load() to retrieve the table data.
+ *
+ * @since 1.0.0
+ *
+ * @param array  $tag_attributes The attributes for the `td` or `th` tag.
+ * @param string $table_id       The current TablePress table ID.
+ * @param string $cell_content   The cell content.
+ * @param int    $row_idx        The row number of the cell.
+ * @param int    $col_idx        The column number of the cell.
+ * @return array The table cell attributes in 'title' => 'value' format.
+ */
+function hrs_filter_tablepress_atts( $tag_attributes, $table_id, $cell_content, $row_idx, $col_idx ) {
+	$table_props = get_transient( 'hrs_tablepress_header_columns' );
+
+	if ( false === $table_props ) {
+		$table       = TablePress::$model_table->load( $table_id );
+		$table_props = array(
+			'header_row' => $table['data'][0],
+		);
+
+		// Save the table properties to a WP transient if they exist.
+		if ( $table_props ) {
+			set_transient( 'hrs_tablepress_header_columns', $table_props, 3600 );
+		}
+	}
+
+	// Get the given cell's column header.
+	$label = $table_props['header_row'][ $col_idx - 1 ];
+
+	// Apply a `data-column` attribute to every cell with a column header.
+	if ( '' !== $label ) {
+		$tag_attributes['data-column'] = esc_attr( $label );
+	}
+
+	return $tag_attributes;
+}
+
+/**
+ * Sets up custom Gravity Forms filters and actions.
+ *
+ * @since 1.1.0
+ *
+ * @return bool True if filters are added.
+ */
+function hrs_gform_setup() {
+	// Check if Gravity Forms exists and is loaded.
+	if ( ! class_exists( 'GFForms' ) ) {
+		return false;
+	}
+
+	/**
+	 * Add filters for the "360 Evaluators Selection" form.
+	 *
+	 * The filter targets list field columns by form, field, column
+	 * combinations in format: `FILTER_NAME_FORMID_FIELDID_COLUMN`.
+	 *
+	 * @uses RGFormsModel()
+	 *
+	 */
+	$form_id = absint( RGFormsModel::get_form_id( '360 Evaluators Selection' ) );
+	if ( $form_id ) {
+		add_filter( 'gform_column_input_content_' . $form_id . '_8_4', 'hrs_filter_evals_reason_column', 10, 6 );
+		add_filter( 'gform_column_input_' . $form_id . '_8_5', 'hrs_filter_evals_rel_column', 10, 5 );
+
+		return true;
+	}
+
+	return false;
+}
+
+/**
+ * Replaces a text field with a textarea field in the target Gravity Forms form.
+ *
+ * This is a callback for the Gravity Forms filter used to modify the HTML of a
+ * targeted List field column input tag.
+ *
+ * @since 1.1.0
+ *
+ * @param string $input      The current HTML content of the List field column.
+ * @param array  $input_info The input array to be filtered.
+ * @param object $field      The current field.
+ * @param string $text       The current column name.
+ * @param string $value      The currently entered/selected value for the column's input.
+ * @param int    $form_id    The ID of the current form.
+ * @return string The HTML content of the List field column.
+ */
+function hrs_filter_evals_reason_column( $input, $input_info, $field, $text, $value, $form_id ) {
+	$input_field_name = 'input_' . $field->id . '[]';
+	$tabindex         = GFCommon::get_tabindex();
+
+	$input = sprintf(
+		'<textarea name="%1$s" %2$s class="textarea medium" cols="40" rows="10">%3$s</textarea>',
+		$input_field_name,
+		$tabindex,
+		$value
+	);
+
+	return $input;
+}
+
+/**
+ * Replaces a text field with a select field in the target Gravity Forms form.
+ *
+ * This is a callback for the Gravity Forms filter used to specify a different
+ * input type for a list field column. Currently supported field types are
+ * “select” (Drop Down) and “text” (Text Field).
+ *
+ * @since 1.1.0
+ *
+ * @param array  $input_info The input info array to be filtered.
+ * @param object $field      The current field.
+ * @param string $column     The current column name.
+ * @param string $value      The currently entered/selected value for the column's input.
+ * @param int    $form_id    The ID of the current form.
+ * @return array The filtered input info array.
+ */
+function hrs_filter_evals_rel_column( $input_info, $field, $column, $value, $form_id ) {
+	$input_info = array(
+		'type'    => 'select',
+		'choices' => array(
+			'',
+			'Peer',
+			'Direct Report',
+			'Student',
+			'Faculty',
+			'Staff',
+			'Regent',
+			'Community Liaison',
+			'Other',
+		),
+	);
+
+	return $input_info;
 }